Serial RX/TX support added + Doc cleanup

+++ JPigpio/src/jpigpio/PigpioSocket.java
@@ -8,11 +8,10 @@ import java.util.ArrayList;
 import jpigpio.impl.CommonPigpio;
 
 /**
- * http://abyz.co.uk/rpi/pigpio/sif.html
- */
-
-/**
- * An implementation of the Pigpio Java interface using sockets to connect to the target pigpio demon.
+ * An implementation of the Pigpio Java interface using sockets to connect to the target pigpio demon
+ * socket interface (see http://abyz.co.uk/rpi/pigpio/sif.html)
+ * <br><br>
+ * See {@link JPigpio JPigpio interface} for full documentation
  *
  */
 public class PigpioSocket extends CommonPigpio {
@@ -59,6 +58,7 @@ public class PigpioSocket extends CommonPigpio {
 	private final int CMD_PRG = 22;			//22 gpio 0 0 -
 	private final int CMD_PFG = 23;			//23 gpio 0 0 -
 	private final int CMD_PRRG = 24;		//24 gpio 0 0 -
+
 	// private final int CMD_HELP = 25;		//25 N/A N/A N/A N/A
 	// private final int CMD_PIGPV = 26;	//26 0 0 0 -
 
@@ -118,20 +118,36 @@ public class PigpioSocket extends CommonPigpio {
 	private final int CMD_SPIW = 74;		// 74 handle 0 X uint8_t data[X]
 	private final int CMD_SPIX = 75;		// 75 handle 0 X uint8_t data[X]
 
-	// CMD_SERO 76 baud flags X uint8_t device[X]
-	// CMD_SERC 77 handle 0 0 -
-	// CMD_SERRB 78 handle 0 0 -
-	// CMD_SERWB 79 handle byte 0 -
-	// CMD_SERR 80 handle count 0 -
-	// CMD_SERW 81 handle 0 X uint8_t data[X]
-	// CMD_SERDA 82 handle 0 0 -
+	private final int CMD_SERO = 76;		// 76 baud flags X uint8_t device[X]
+	private final int CMD_SERC = 77;		// 77 handle 0 0 -
+	private final int CMD_SERRB = 78;		// 78 handle 0 0 -
+	private final int CMD_SERWB = 79;		// 79 handle byte 0 -
+	private final int CMD_SERR = 80;		// 80 handle count 0 -
+	private final int CMD_SERW = 81;		// 81 handle 0 X uint8_t data[X]
+	private final int CMD_SERDA = 82;		// 82 handle 0 0 -
 
 	private final int CMD_GDC = 83;			// 83 gpio 0 0 -
+
 	// CMD_GPW 84 gpio 0 0 -
 	// CMD_HC 85 gpio frequency 0 -
 	// CMD_HP 86 gpio frequency 4 uint32_t dutycycle
 	// CMD_CF1 87 arg1 arg2 X uint8_t argx[X]
 	// CMD_CF2 88 arg1 retMax X uint8_t argx[X]
+	// CMD_BI2CC	89	sda	0	0	-
+	// CMD_BI2CO	90	sda	scl	4	uint32_t baud
+	// CMD_BI2CZ	91	sda	0	X	uint8_t data[X]
+	// CMD_I2CZ	92	handle	0	X	uint8_t data[X]
+	// CMD_WVCHA	93	0	0	X	uint8_t data[X]
+	// CMD_SLRI	94	gpio	invert	0	-
+	// CMD_CGI	95	0	0	0	-
+	// CMD_CSI	96	config	0	0	-
+	// CMD_FG	97	gpio	steady	0	-
+	// CMD_FN	98	gpio	steady	4	uint32_t active
+	// CMD_NOIB	99	0	0	0	-
+	// CMD_WVTXM	100	wave_id	mode	0	-
+	// CMD_WVTAT	101	-	-	0	-
+
+
 
 	private final int CMD_NOIB = 99;		//99 0 0 0 -
 
@@ -249,6 +265,7 @@ public class PigpioSocket extends CommonPigpio {
 			int gpio = 0;
 
 			try {
+				// read GPIO status for GPIOs in bank 1 (gpio 0-31)
 				int lastLevel = slPiCmd.sendCmd(CMD_BR1, 0, 0);
 
 				// loop until stop signal is received
@@ -280,7 +297,7 @@ public class PigpioSocket extends CommonPigpio {
 									newLevel = 1;
 								//System.out.println("#3 "+changed+" : "+Integer.toBinaryString(cb.bit)+" : "+(cb.bit & changed));
 								if ((cb.edge ^ newLevel) != 0)
-									cb.processNotification(cb.gpio, newLevel, tick);
+									cb.alert(cb.gpio, newLevel, tick);
 
 							}
 						}
@@ -290,7 +307,7 @@ public class PigpioSocket extends CommonPigpio {
 							gpio = flags & PI_NTFY_FLAGS_WDOG;
 							for (GPIOListener cb : gpioListeners)
 								if (cb.gpio == gpio)
-									cb.processNotification(cb.gpio, PI_TIMEOUT, tick);
+									cb.alert(cb.gpio, PI_TIMEOUT, tick);
 					}
 				}
 
@@ -318,10 +335,9 @@ public class PigpioSocket extends CommonPigpio {
 	/**
 	 * The constructor of the class.
 	 * 
-	 * @param host
-	 *            The address of the pigpio daemon.
-	 * @param port
-	 *            The port of the pigpio daemon.
+	 * @param host The host name or ip address of the pigpio daemon.
+	 * @param port The port of the pigpio daemon.
+	 * @throws  PigpioException if not able to initialize/connect to pigpiod
 	 */
 	public PigpioSocket(String host, int port) throws PigpioException {
 		this.host = host;
@@ -329,11 +345,6 @@ public class PigpioSocket extends CommonPigpio {
 		gpioInitialize();
 	}
 
-	/**
-	 * Initialize
-	 * 
-	 * @return
-	 */
 	@Override
 	public void gpioInitialize() throws PigpioException {
 		try {
@@ -355,9 +366,15 @@ public class PigpioSocket extends CommonPigpio {
 	public void gpioTerminate() throws PigpioException {
 		try {
 			// stop listener thread
-			router.terminate();
+			if (router != null) {
+				router.terminate();
+				router = null;
+			}
 			// stop command socket to pigpio
-			slCmd.terminate();
+			if (slCmd != null) {
+				slCmd.terminate();
+				slCmd = null;
+			}
 		} catch (Exception e) {
 			throw new PigpioException("gpioTerminate", e);
 		}
@@ -404,9 +421,8 @@ public class PigpioSocket extends CommonPigpio {
 	public boolean gpioRead(int pin) throws PigpioException {
 		try {
 			int rc = slCmd.sendCmd(CMD_READ, pin, 0);
-			if (rc < 0) {
+			if (rc < 0)
 				throw new PigpioException(rc);
-			}
 			return rc != 0;
 		} catch (IOException e) {
 			throw new PigpioException("gpioRead", e);
@@ -417,9 +433,8 @@ public class PigpioSocket extends CommonPigpio {
 	public void gpioWrite(int pin, boolean value) throws PigpioException {
 		try {
 			int rc = slCmd.sendCmd(CMD_WRITE, pin, value?1:0);
-			if (rc < 0) {
+			if (rc < 0)
 				throw new PigpioException(rc);
-			}
 		} catch (IOException e) {
 			throw new PigpioException("gpioWrite", e);
 		}
@@ -430,25 +445,33 @@ public class PigpioSocket extends CommonPigpio {
 	@Override
 	public int notifyOpen() throws PigpioException {
 		try {
-			return slCmd.sendCmd(CMD_NO, 0, 0);
+			int rc = slCmd.sendCmd(CMD_NO, 0, 0);
+			if (rc < 0)
+				throw new PigpioException(rc);
+			return rc;
 		} catch (IOException e) {
 			throw new PigpioException("notifyOpen", e);
 		}
 	}
 
 	@Override
-	public int notifyBegin(int handle, int bits) throws PigpioException{
+	public void notifyBegin(int handle, int bits) throws PigpioException{
 		try {
-			return slCmd.sendCmd(CMD_NB, handle, bits);
+			int rc = slCmd.sendCmd(CMD_NB, handle, bits);
+			if (rc < 0)
+				throw new PigpioException(rc);
 		} catch (IOException e) {
 			throw new PigpioException("notifyBegin", e);
 		}
 	}
 
 	@Override
-	public int notifyPause(int handle) throws PigpioException {
+	public void notifyPause(int handle) throws PigpioException {
 		try {
-			return slCmd.sendCmd(CMD_NP, handle, 0);
+			int rc = slCmd.sendCmd(CMD_NP, handle, 0);
+			if (rc < 0)
+				throw new PigpioException(rc);
+
 		} catch (IOException e) {
 			throw new PigpioException("notifyPause", e);
 		}
@@ -456,9 +479,12 @@ public class PigpioSocket extends CommonPigpio {
 	}
 
 	@Override
-	public int notifyClose(int handle) throws PigpioException{
+	public void notifyClose(int handle) throws PigpioException{
 		try {
-			return slCmd.sendCmd(CMD_NC, handle, 0);
+			int rc = slCmd.sendCmd(CMD_NC, handle, 0);
+			if (rc < 0)
+				throw new PigpioException(rc);
+
 		} catch (IOException e) {
 			throw new PigpioException("notifyClose", e);
 		}
@@ -466,9 +492,12 @@ public class PigpioSocket extends CommonPigpio {
 	}
 
 	@Override
-	public int setWatchdog(int userGpio, int timeout) throws PigpioException{
+	public void setWatchdog(int userGpio, int timeout) throws PigpioException{
 		try {
-			return slCmd.sendCmd(CMD_WDOG, userGpio, timeout);
+			int rc =  slCmd.sendCmd(CMD_WDOG, userGpio, timeout);
+			if (rc < 0)
+				throw new PigpioException(rc);
+
 		} catch (IOException e) {
 			throw new PigpioException("setWatchdog", e);
 		}
@@ -476,15 +505,13 @@ public class PigpioSocket extends CommonPigpio {
 
 	// ################ WAVES
 
-	/**
-	 * This function clears all waveforms and any data added by calls to the wave_add_* functions.
-	 *
-	 * @return The return code from close.
-	 */
 	@Override
-	public int waveClear() throws PigpioException {
+	public void waveClear() throws PigpioException {
 		try {
-			return slCmd.sendCmd(CMD_WVCLR, 0, 0);
+			int rc = slCmd.sendCmd(CMD_WVCLR, 0, 0);
+			if (rc < 0)
+				throw new PigpioException(rc);
+
 		} catch (IOException e) {
 			throw new PigpioException("waveClear", e);
 		}
@@ -501,7 +528,7 @@ public class PigpioSocket extends CommonPigpio {
 		// III on/off/delay * pulses
 
         ByteBuffer bb;
-        int ret;
+        int rc;
 
 		if (pulses == null || pulses.size() == 0)
 			return 0;
@@ -512,8 +539,11 @@ public class PigpioSocket extends CommonPigpio {
 			for (Pulse p:pulses)
                 bb.putInt(p.gpioOn).putInt(p.gpioOff).putInt(p.delay);
 
-            ret = slCmd.sendCmd(CMD_WVAG,0,0,pulses.size()*12,bb.array());
-			return ret;
+            rc = slCmd.sendCmd(CMD_WVAG,0,0,pulses.size()*12,bb.array());
+			if (rc < 0)
+				throw new PigpioException(rc);
+
+			return rc;
 
 		} catch (IOException e) {
 			throw new PigpioException("waveAddGeneric", e);
@@ -545,77 +575,50 @@ public class PigpioSocket extends CommonPigpio {
             bb.order(ByteOrder.LITTLE_ENDIAN);
             bb.putInt(bbBits).putInt(bbStop).putInt(offset);
 
-			return slCmd.sendCmd(CMD_WVAS, userGpio, baud, data.length + 12, bb.array());
+			int rc = slCmd.sendCmd(CMD_WVAS, userGpio, baud, data.length + 12, bb.array());
+			if (rc < 0)
+				throw new PigpioException(rc);
+
+			return rc;
 		} catch (IOException e) {
 			throw new PigpioException("waveAddSerial", e);
 		}
 
 	}
 
-	/**
-	 * Starts a new empty waveform.
-	 *
-	 * You would not normally need to call this function as it is
-	 * automatically called after a waveform is created with the
-	 * [*wave_create*] function.
-	 *
-	 * ...
-	 * pi.wave_add_new()
-	 * ...
-	 *
-	 * @return The return code from add new.
-	 */
+
 	@Override
-	public int waveAddNew() throws PigpioException {
+	public void waveAddNew() throws PigpioException {
 		try {
-			return slCmd.sendCmd(CMD_WVNEW, 0, 0);
+			int rc = slCmd.sendCmd(CMD_WVNEW, 0, 0);
+			if (rc < 0)
+				throw new PigpioException(rc);
+
 		} catch (IOException e) {
 			throw new PigpioException("waveAddNew", e);
 		}
 	} // waveAddNew
 
-	/**
-	 * Returns 1 if a waveform is currently being transmitted,
-	 * otherwise 0.
-	 *
-	 * ...
-	 * pi.wave_send_once(0) # send first waveform
-	 *
-	 * while pi.wave_tx_busy(): # wait for waveform to be sent
-	 * time.sleep(0.1)
-	 *
-	 * pi.wave_send_once(1) # send next waveform
-	 * ...
-	 * @return The return code from wave_tx_busy.
-	 */
+
 	@Override
 	public boolean waveTxBusy() throws PigpioException {
 		try {
-			return (slCmd.sendCmd(CMD_WVBSY, 0, 0) != 0);
+			int rc = slCmd.sendCmd(CMD_WVBSY, 0, 0);
+			if (rc < 0)
+				throw new PigpioException(rc);
+			return rc != 0;
 		} catch (IOException e) {
 			throw new PigpioException("waveTxBusy", e);
 		}
 	} // waveTxBusy
 
-	/**
-	 * Stops the transmission of the current waveform.
-	 *
-	 * This function is intended to stop a waveform started with
-	 * wave_send_repeat.
-	 *
-	 * ...
-	 * pi.wave_send_repeat(3)
-	 *
-	 * time.sleep(5)
-	 *
-	 * pi.wave_tx_stop()
-	 * ...
-	 * @return The return code from wave_tx_stop.
-	 */
 	@Override
 	public int waveTxStop() throws PigpioException {
 		try {
-			return slCmd.sendCmd(CMD_WVHLT, 0, 0);
+			int rc = slCmd.sendCmd(CMD_WVHLT, 0, 0);
+			if (rc < 0)
+				throw new PigpioException(rc);
+			return rc;
 		} catch (IOException e) {
 			throw new PigpioException("waveTxStop", e);
 		}
@@ -624,16 +627,22 @@ public class PigpioSocket extends CommonPigpio {
 	@Override
 	public int waveCreate() throws PigpioException {
 		try {
-			return slCmd.sendCmd(CMD_WVCRE, 0, 0);
+			int rc = slCmd.sendCmd(CMD_WVCRE, 0, 0);
+			if (rc < 0)
+				throw new PigpioException(rc);
+			return rc;
 		} catch (IOException e) {
 			throw new PigpioException("waveCreate", e);
 		}
 	}
 
 	@Override
-	public int waveDelete(int waveId) throws PigpioException{
+	public void waveDelete(int waveId) throws PigpioException{
 		try {
-			return slCmd.sendCmd(CMD_WVDEL, 0, 0);
+			int rc = slCmd.sendCmd(CMD_WVDEL, 0, 0);
+			if (rc < 0)
+				throw new PigpioException(rc);
+
 		} catch (IOException e) {
 			throw new PigpioException("waveDelete", e);
 		}
@@ -642,7 +651,10 @@ public class PigpioSocket extends CommonPigpio {
 	@Override
 	public int waveSendOnce(int waveId) throws PigpioException {
 		try {
-			return slCmd.sendCmd(CMD_WVTX, waveId, 0);
+			int rc = slCmd.sendCmd(CMD_WVTX, waveId, 0);
+			if (rc < 0)
+				throw new PigpioException(rc);
+			return rc;
 		} catch (IOException e) {
 			throw new PigpioException("waveSendOnce", e);
 		}
@@ -651,7 +663,11 @@ public class PigpioSocket extends CommonPigpio {
 	@Override
 	public int waveSendRepeat(int waveId) throws PigpioException {
 		try {
-			return slCmd.sendCmd(CMD_WVTXR, waveId, 0);
+			int rc = slCmd.sendCmd(CMD_WVTXR, waveId, 0);
+			if (rc < 0)
+				throw new PigpioException(rc);
+			return rc;
+
 		} catch (IOException e) {
 			throw new PigpioException("waveSendRepeat", e);
 		}
@@ -682,7 +698,6 @@ public class PigpioSocket extends CommonPigpio {
 	 * 
 	 * @param handle
 	 *            The handle of the previously opened i2c
-	 * @return The return code from the close.
 	 */
 	@Override
 	public void i2cClose(int handle) throws PigpioException {
@@ -792,23 +807,24 @@ public class PigpioSocket extends CommonPigpio {
 		}
 	} // End of gpioServo
 
+	@Override
 	public void setServoPulseWidth(int gpio, int pulseWidth) throws PigpioException {
 		gpioServo(gpio, pulseWidth);
 	}
 
+	@Override
 	public int getServoPulseWidth(int gpio) throws PigpioException {
 		throw new NotImplementedException();
 	}
 
-	/**
-	 * 
-	 * @param pin
-	 * @param alert
-	 * @throws PigpioException
-	 */
 	@Override
-	public void gpioSetAlertFunc(int pin, Alert alert) throws PigpioException {
-		throw new NotImplementedException();
+	public void gpioSetAlertFunc(int pin, Alert gpioAlert) throws PigpioException {
+		router.addListener(new GPIOListener(pin, PI_EITHER_EDGE) {
+			@Override
+			public void alert(int gpio, int level, long tick) {
+				gpioAlert.alert(gpio, level, tick);
+			}
+		});
 	} // End of gpioSetAlertFunc
 
 	@Override
@@ -818,14 +834,6 @@ public class PigpioSocket extends CommonPigpio {
 
 	// ############### SPI
 	
-	/**
-	 *
-	 * @param spiChannel The channel to open.
-	 * @param spiBaudRate The baud rate for transmition and receiption
-	 * @param flags Control flags
-	 * @return A handle used in subsequent SPI API calls
-	 * @throws PigpioException
-	 */
 	@Override
 	public int spiOpen(int spiChannel, int spiBaudRate, int flags) throws PigpioException {
 		int rc = 0;
@@ -846,11 +854,6 @@ public class PigpioSocket extends CommonPigpio {
 	}
 
 
-	/**
-	 * Close an SPI connection previously created with spiOpen().
-	 * @param handle The handle to be closed.
-	 * @throws PigpioException
-	 */
 	@Override
 	public void spiClose(int handle) throws PigpioException {
 		try {
@@ -861,13 +864,6 @@ public class PigpioSocket extends CommonPigpio {
 	}
 
 
-	/**
-	 * Read data from SPI.
-	 * @param handle The handle from which to read.
-	 * @param data An array into which to read data.
-	 * @return The number of bytes actually read.
-	 * @throws PigpioException
-	 */
 	@Override
 	public int spiRead(int handle, byte[] data) throws PigpioException {
 		int rc = 0;
@@ -885,13 +881,6 @@ public class PigpioSocket extends CommonPigpio {
 		}
 	}
 
-	/**
-	 * Write data to SPI.
-	 * @param handle The handle into which to write.
-	 * @param data An array of data to write to SPI.
-	 * @return The number of bytes actually written
-	 * @throws PigpioException
-	 */
 	@Override
 	public int spiWrite(int handle, byte[] data) throws PigpioException {
 		int rc = 0;
@@ -907,15 +896,7 @@ public class PigpioSocket extends CommonPigpio {
 		return rc;
 	}
 
-	/**
-	 * Write data to SPI and in parallel, read responses.  The size of the txData and rxData arrays must
-	 * be the same.
-	 * @param handle The handle into which to write.
-	 * @param txData An array of data to write.
-	 * @param rxData An array of data to read.
-	 * @return The number of bytes actually transferred.
-	 * @throws PigpioException
-	 */
+
 	@Override
 	public int spiXfer(int handle, byte[] txData, byte[] rxData) throws PigpioException {
 		int rc = 0;
@@ -936,8 +917,7 @@ public class PigpioSocket extends CommonPigpio {
 
 	// ######################## PWM
 
-	// ############### PWM
-
+	@Override
 	public void setPWMDutycycle(int gpio, int dutycycle) throws PigpioException {
 		try {
 			slCmd.sendCmd(CMD_PWM, gpio, dutycycle);
@@ -947,7 +927,7 @@ public class PigpioSocket extends CommonPigpio {
 
 	}
 
-
+	@Override
 	public int getPWMDutycycle(int gpio) throws PigpioException {
 		int rc = 0;
 		try {
@@ -963,7 +943,7 @@ public class PigpioSocket extends CommonPigpio {
 
 	}
 
-
+	@Override
 	public void setPWMRange(int gpio, int range) throws PigpioException {
 		try {
 			slCmd.sendCmd(CMD_PRS, gpio, range);
@@ -973,7 +953,7 @@ public class PigpioSocket extends CommonPigpio {
 
 	}
 
-
+	@Override
 	public int getPWMRange(int gpio) throws PigpioException {
 		int rc = 0;
 		try {
@@ -988,7 +968,7 @@ public class PigpioSocket extends CommonPigpio {
 		return rc;
 	}
 
-
+	@Override
 	public int getPWMRealRange(int gpio) throws PigpioException {
 		int rc = 0;
 		try {
@@ -1003,7 +983,7 @@ public class PigpioSocket extends CommonPigpio {
 		return rc;
 	}
 
-
+	@Override
 	public int setPWMFrequency(int gpio, int frequency) throws PigpioException {
 
 		int rc = 0;
@@ -1020,7 +1000,7 @@ public class PigpioSocket extends CommonPigpio {
 
 	}
 
-
+	@Override
 	public int getPWMFrequency(int gpio) throws PigpioException {
 		int rc = 0;
 		try {
@@ -1035,6 +1015,104 @@ public class PigpioSocket extends CommonPigpio {
 		return rc;
 	}
 
+
+	// ################ SERIAL
+	@Override
+	public int serialOpen(String tty, int baudRate, int flags) throws PigpioException {
+		int rc = 0;
+
+		try {
+			rc = slCmd.sendCmd(CMD_SERO, baudRate, flags, tty.length(), tty.getBytes());
+			if (rc < 0)
+				throw new PigpioException(rc);
+
+		} catch (IOException e) {
+			throw new PigpioException("serialOpen failed", e);
+		}
+
+		return rc;
+
+	}
+
+	@Override
+	public void serialClose(int handle) throws PigpioException {
+		try {
+			slCmd.sendCmd(CMD_SERC, handle, 0);
+		} catch (IOException e) {
+			throw new PigpioException("serialClose failed",e);
+		}
+	}
+
+	@Override
+	public byte serialReadByte(int handle) throws PigpioException {
+		int rc = 0;
+
+		try {
+			rc = slCmd.sendCmd(CMD_SERRB, handle, 0);
+			if (rc < 0)
+				throw new PigpioException(rc);
+
+		} catch (IOException e) {
+			throw new PigpioException("serialReadByte failed", e);
+		}
+
+		return (byte)rc;
+	}
+
+	@Override
+	public void serialWriteByte(int handle, byte data) throws PigpioException {
+		try {
+			slCmd.sendCmd(CMD_SERWB, handle, data);
+		} catch (IOException e) {
+			throw new PigpioException("serialWriteByte failed",e);
+		}
+	}
+
+	@Override
+	public byte[] serialRead(int handle, int count) throws PigpioException {
+		byte[] data = new byte[1];
+		int rc = 0;
+
+		try {
+			rc = slCmd.sendCmd(CMD_SERR, handle, count );
+			if (rc < 0)
+				throw new PigpioException(rc);
+			if (rc > 0) {
+				data = new byte[rc];
+				slCmd.readBytes(data);
+			}
+
+		} catch (IOException e) {
+			throw new PigpioException("serialRead failed", e);
+		}
+
+		return data;
+	}
+
+	@Override
+	public void serialWrite(int handle, byte[] data) throws PigpioException {
+		try {
+			slCmd.sendCmd(CMD_SERW, handle, 0, data.length, data);
+		} catch (IOException e) {
+			throw new PigpioException("serialWrite failed",e);
+		}
+	}
+
+	@Override
+	public int serialDataAvailable(int handle) throws PigpioException {
+		int rc = 0;
+
+		try {
+			rc = slCmd.sendCmd(CMD_SERDA, handle,0);
+			if (rc < 0)
+				throw new PigpioException(rc);
+
+		} catch (IOException e) {
+			throw new PigpioException("serialDataAvailable failed", e);
+		}
+
+		return rc;
+	}
 
 	// ########################
